Add package manager
