+++ src/js/api.ts
@@ -1,13 +1,13 @@
 import { ffi } from "./ffi";
 import { CanvasInterface, canvas } from "./canvas";
 
-import { PackageData, loadPackage, loadedPackages } from "./load-package";
-import { type PyProxy, type PyDict } from "generated/pyproxy";
+import { loadPackage, loadedPackages } from "./package-manager";
+import { type PyProxy, type PyDict, isPyProxy } from "generated/pyproxy";
 import { loadBinaryFile, nodeFSMod } from "./compat";
 import { version } from "./version";
 import { setStdin, setStdout, setStderr } from "./streams";
 import { scheduleCallback } from "./scheduler";
-import { TypedArray } from "./types";
+import { TypedArray, PackageData } from "./types";
 import { IN_NODE, detectEnvironment } from "./environments";
 // @ts-ignore
 import LiteralMap from "./common/literal-map";
@@ -77,6 +77,8 @@ if (AbortSignal.any) {
 }
 
 API.LiteralMap = LiteralMap;
+
+API.isPyProxy = isPyProxy;
 
 function ensureMountPathExists(path: string): void {
   Module.FS.mkdirTree(path);
