--- src/js/common/memoize.ts
@@ -1,18 +1,0 @@
-/**
- * @param fn A function to be memoized.
- * @returns
- * @private
- */
-export const memoize = (fn: CallableFunction) => {
-  let cache: any = {};
-  return (...args: any) => {
-    let n = args[0];
-    if (n in cache) {
-      return cache[n];
-    } else {
-      let result = fn(n);
-      cache[n] = result;
-      return result;
-    }
-  };
-};
