+++ src/js/pyodide.ts
@@ -13,9 +13,9 @@ import { createSettings } from "./emscripten-settings";
 import { version } from "./version";
 
 import type { PyodideInterface } from "./api.js";
-import type { TypedArray, Module } from "./types";
+import type { TypedArray, Module, PackageData } from "./types";
 import type { EmscriptenSettings } from "./emscripten-settings";
-import type { PackageData } from "./load-package";
+import { initializePackageIndex } from "./package-manager";
 import { SnapshotConfig } from "./snapshot";
 export type { PyodideInterface, TypedArray };
 
@@ -231,6 +231,7 @@ export async function loadPyodide(
   const emscriptenSettings = createSettings(config);
   const API = emscriptenSettings.API;
   API.lockFilePromise = loadLockFile(config.lockFileURL);
+  API.packageIndexReady = initializePackageIndex(API.lockFilePromise);
 
   // If the pyodide.asm.js script has been imported, we can skip the dynamic import
   // Users can then do a static import of the script in environments where
